+++ programs/onlybags-staking/src/account_data/user_info.rs
@@ -6,7 +6,7 @@ use anchor_lang::prelude::*;
 pub struct UserInfo {
   pub bump: u8,
   pub staked_amount: u64,
-  pub reward_debt: u128,
+  pub reward_debt: u64,
   pub acc_claim: u64,
   pub total_claimed: u64,
 }
